+++ bin/console.php
@@ -1,9 +1,0 @@
-#!/usr/bin/env php
-<?php
-
-require __DIR__ . '/../vendor/autoload.php';
-
-use Application\Component\Console\Application;
-
-$application = new Application();
-$application->run();