+++ src/Application/Component/Console/Command/ImportCommand/ImportCommand.php
@@ -1,25 +1,21 @@
 <?php
 
-namespace Application\Component\Console\Command;
+namespace Application\Component\Console\Command\ImportCommand;
 
 use Application\Component\Filesystem\Filesystem;
-use Application\Component\Finder\Finder;
 use Application\Exception\InvalidArgumentException;
 use Application\Exception\RuntimeException;
 use NumberFormatter;
-use Ramsey\Uuid\Uuid;
 use Symfony\Component\Console\Helper\ProgressBar;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class FortuneImportCommand extends AbstractCommand
+class ImportCommand extends AbstractCommand
 {
-    const FORTUNES_PER_FILE = 50;
-
     protected function configure()
     {
-        $this->setName('fortune-import');
+        $this->setName('import');
 
         $this->setDescription('Import fortunes from JSON files');
 
@@ -101,57 +97,5 @@ class FortuneImportCommand extends AbstractCommand
                                  (1 === $curFortunesCount) ? 'fortune' : 'fortunes'));
 
         return $this;
-    }
-
-    private function getNewFortunes($inputPath)
-    {
-        $ret = [];
-
-        $finder = new Finder();
-
-        foreach ($finder->json($inputPath) as $fileInfo) {
-            $json  = file_get_contents($fileInfo->getPathname());
-            $array = json_decode($json, true);
-            foreach ($array as $record) {
-                $quote  = $this->filter($record['quoteText']);
-                $author = $this->filter($record['quoteAuthor']);
-                $uuid   = $this->uuid($quote);
-                if (array_key_exists($uuid, $ret)) {
-                    continue;
-                }
-                if (empty($quote)) {
-                    continue;
-                }
-                if (empty($author)) {
-                    $author = 'Unknown';
-                }
-                $ret[$uuid] = [
-                    $quote,
-                    $author,
-                ];
-            }
-        }
-
-        shuffle($ret);
-
-        return $ret;
-    }
-
-    private function filter($string)
-    {
-        $string = str_replace('’', "'", $string);
-        $string = trim($string);
-
-        return $string;
-    }
-
-    private function uuid($quote)
-    {
-        $name = strtolower($quote);
-        $name = preg_replace('/[^a-z]/', null, $name);
-
-        $uuid5 = Uuid::uuid5(Uuid::NIL, $name);
-
-        return strtolower($uuid5->toString());
     }
 }