+++ src/Application/Component/Console/Command/IndexCommand/AbstractCommand.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Application\Component\Console\Command\IndexCommand;
+
+use Application\Component\Console\Command\AbstractCommand as ParentCommand;
+
+abstract class AbstractCommand extends ParentCommand
+{
+}