+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## dev-master
 
-* Added command `fortune-statistics`.
-* Added command `fortune-index`.
+* Added `--short` option to command `fortune`.
+* Added `--long` option to command `fortune`.
+* Added command `statistics`.
+* Added command `index`.
 * Added `--length` option to command `fortune`. 
 * Added `--author` option to command `fortune`. 
 * Set the default `--wordwrap` og command `fortune` to the terminal window width minus one.